Created seperate class file for SongMap.cs
