+++ SongMap.cs
@@ -1,0 +1,14 @@
+﻿using Converters;
+using CsvHelper.Configuration;
+public class SongMap : ClassMap<Song>
+{
+    public SongMap()
+    {
+        Map(m => m.Name);
+        Map(m => m.Artist);
+        Map(m => m.Composer);
+        Map(m => m.Genre);
+        Map(m => m.Year).TypeConverter<CustomDateYearConverter>();
+        Map(m => m.Plays).TypeConverter<CustomIntConverter>();
+    }
+}
